fixed issues in book return and recursively called returnBoook and issueBook (#18)

+++ src/transaction-management/transaction.interaction.ts
@@ -23,6 +23,7 @@ import { MemberRepository } from "../member-management/member.repository";
 import { IBook } from "../models/book.schema";
 import { IMember } from "../models/member.schema";
 import { IPageRequest } from "../core/pagination";
+import { error } from "console";
 
 const menu = new Menu([
   { key: "1", label: "Issue Book" },
@@ -244,29 +245,40 @@ async function issueBook(
   bookRepo: BookRepository,
   memberRepo: MemberRepository
 ) {
-  try {
-    const bookIssueData: ITransactionBase = await getIssueInput(
-      bookRepo,
-      memberRepo
-    );
-    const createdTransaction: ITransaction = await repo.create(bookIssueData);
-    if (createdTransaction) {
-      const book = await bookRepo.getById(createdTransaction.bookId);
-      const member = await memberRepo.getById(createdTransaction.memberId);
-      if (book && member) {
-        printHint("\nBook issue details:");
-        console.table({
-          "Book Title": book.title,
-          "Member Name": member.name,
-          "Issue Date": createdTransaction.dateOfIssue.toDateString(),
-          "Return Date": createdTransaction.dueDate.toDateString(),
-        });
+  let issueMoreBooks = true;
+
+  while (issueMoreBooks) {
+    try {
+      const bookIssueData: ITransactionBase = await getIssueInput(
+        bookRepo,
+        memberRepo
+      );
+
+      const createdTransaction: ITransaction = await repo.create(bookIssueData);
+      if (createdTransaction) {
+        const book = await bookRepo.getById(createdTransaction.bookId);
+        const member = await memberRepo.getById(createdTransaction.memberId);
+        if (book && member) {
+          printHint("\nBook issue details:");
+          console.table({
+            "Book Title": book.title,
+            "Member Name": member.name,
+            "Issue Date": createdTransaction.dateOfIssue.toDateString(),
+            "Return Date": createdTransaction.dueDate.toDateString(),
+          });
+        }
+        printResult(`Book issued successfully.`);
       }
-      printResult(`Book issued successfully.`);
+    } catch (error: unknown) {
+      if (error instanceof Error) console.log(error.message);
     }
-  } catch (error: unknown) {
-    if (error instanceof Error) console.log(error.message);
+
+    const continueIssuing = await readLine(
+      "Do you want to issue another book? (y/n): "
+    );
+    issueMoreBooks = continueIssuing.toLowerCase() === "y";
   }
+
   printHint(`Press ${enterButton} to continue`);
   await readLine("");
   return;
@@ -277,24 +289,47 @@ async function returnBook(
   bookRepo: BookRepository,
   memberRepo: MemberRepository
 ) {
-  try {
-    const bookReturnData: ITransactionBase = await getReturnInput();
-    const id = bookReturnData.bookId;
-    const deletedTransaction: ITransaction | null = await repo.delete(id);
-    if (deletedTransaction) {
-      const book = await bookRepo.getById(deletedTransaction.bookId);
-      const member = await memberRepo.getById(deletedTransaction.memberId);
-      if (book && member) {
-        console.table({
-          "Book Title": book.title,
-          "Member Name": member.name,
-        });
+  let returnMoreBooks = true;
+  while (returnMoreBooks) {
+    try {
+      const bookReturnData: ITransactionBase = await getReturnInput();
+      const transactions = await repo.list(`${bookReturnData.bookId}`);
+      const transaction = transactions.find((item) => {
+        return (
+          item.bookId === bookReturnData.bookId &&
+          item.memberId === bookReturnData.memberId &&
+          item.bookStatus === "issued"
+        );
+      });
+
+      if (transaction) {
+        const id = transaction.id;
+        const deletedTransaction: ITransaction | null = await repo.delete(id);
+        if (deletedTransaction) {
+          const book = await bookRepo.getById(deletedTransaction.bookId);
+          const member = await memberRepo.getById(deletedTransaction.memberId);
+          if (book && member) {
+            printHint("\nBook return details:");
+            console.table({
+              "Book Title": book.title,
+              "Member Name": member.name,
+            });
+          }
+          printResult(`Book returned successfully.`);
+        }
+      } else {
+        throw new Error("Book already returned or Transaction does not exist");
       }
-      printResult(`Book returned successfully.`);
+    } catch (error: unknown) {
+      if (error instanceof Error) printError(error.message);
     }
-  } catch (error: unknown) {
-    if (error instanceof Error) console.log(error.message);
+
+    const continueReturning = await readLine(
+      "Do you want to return another book? (y/n): "
+    );
+    returnMoreBooks = continueReturning.toLowerCase() === "y";
   }
+
   printHint(`Press ${enterButton} to continue`);
   await readLine("");
   return;
